refactor: extract MCP functionality into dedicated MCPManager
- Move MCP-related code from Agent class to new MCPManager class
- Reduce Agent class from 910 to 776 lines (134 lines removed)
- Place MCPManager in mcp/ module for better organization
- Remove redundant public methods, expose mcp_manager directly
- Maintain backward compatibility with existing API
- Clean separation of concerns between Agent and MCP functionality
