+++ agents-core/stream_agents/core/agents/agents.py
@@ -6,7 +6,6 @@ from uuid import uuid4
 
 import aiortc
 from aiortc import VideoStreamTrack
-from aiortc.contrib.media import MediaRelay
 
 from ..edge.types import Participant, PcmData, Connection, TrackType, User
 from ..llm.events import RealtimePartialTranscriptEvent
@@ -19,8 +18,7 @@ from ..llm.events import RealtimeTranscriptEvent, LLMResponseEvent
 from ..stt.events import STTTranscriptEvent, STTPartialTranscriptEvent
 from ..vad.events import VADAudioEvent
 from getstream.video.rtc import Call
-from ..mcp import MCPBaseServer
-from ..mcp.tool_converter import MCPToolConverter
+from ..mcp import MCPBaseServer, MCPManager
 
 
 from .conversation import StreamHandle, Message, Conversation
@@ -44,7 +42,7 @@ logger = logging.getLogger(__name__)
 def _log_task_exception(task: asyncio.Task):
     try:
         task.result()
-    except Exception as e:
+    except Exception:
         logger.exception("Error in background task")
 
 class Agent:
@@ -112,6 +110,9 @@ class Agent:
         self.vad = vad
         self.processors = processors or []
         self.mcp_servers = mcp_servers or []
+        
+        # Initialize MCP manager if servers are provided
+        self.mcp_manager = MCPManager(self.mcp_servers, self.llm, self.logger) if self.mcp_servers else None
 
         # we sync the user talking and the agent responses to the conversation
         # because we want to support streaming responses and can have delta updates for both
@@ -183,7 +184,9 @@ class Agent:
         self.call = call
         self.conversation = None
 
-        await self._connect_mcp_servers()
+        # Connect to MCP servers if manager is available
+        if self.mcp_manager:
+            await self.mcp_manager.connect_all()
 
         # Setup chat and connect it to transcript events
         self.conversation = self.edge.create_conversation(
@@ -259,7 +262,8 @@ class Agent:
         self._agent_conversation_handle = None
 
         # Disconnect from MCP servers
-        await self._disconnect_mcp_servers()
+        if self.mcp_manager:
+            await self.mcp_manager.disconnect_all()
 
         # Close Realtime connection
         if self._realtime_connection:
@@ -418,14 +422,6 @@ class Agent:
             )
             self.logger.error(f"Error in agent say: {e}")
 
-    def _setup_mcp_servers(self):
-        """Set up MCP servers if any are configured."""
-        if self.mcp_servers:
-            self.logger.info(f"🔌 Setting up {len(self.mcp_servers)} MCP server(s)")
-            for i, server in enumerate(self.mcp_servers):
-                self.logger.info(f"  {i + 1}. {server.__class__.__name__}")
-        else:
-            self.logger.debug("No MCP servers configured")
 
     def _setup_turn_detection(self):
         if self.turn_detection:
@@ -560,7 +556,7 @@ class Agent:
             except asyncio.CancelledError:
                 return
 
-            except Exception as e:
+            except Exception:
                 raise
 
         # Cleanup and logging
@@ -769,128 +765,6 @@ class Agent:
             self._video_track = video_publisher.create_video_track()
             self.logger.info("🎥 Video track initialized from video publisher")
 
-    async def _connect_mcp_servers(self):
-        """Connect to all configured MCP servers and register their tools."""
-        if not self.mcp_servers:
-            return
-
-        self.logger.info(f"🔌 Connecting to {len(self.mcp_servers)} MCP server(s)")
-
-        for i, server in enumerate(self.mcp_servers):
-            try:
-                self.logger.info(
-                    f"  Connecting to MCP server {i + 1}/{len(self.mcp_servers)}: {server.__class__.__name__}"
-                )
-                await server.connect()
-                self.logger.info(
-                    f"  ✅ Connected to MCP server {i + 1}/{len(self.mcp_servers)}"
-                )
-
-                # Register MCP tools with the LLM's function registry
-                await self._register_mcp_tools(i, server)
-
-            except Exception as e:
-                self.logger.error(
-                    f"  ❌ Failed to connect to MCP server {i + 1}/{len(self.mcp_servers)}: {e}"
-                )
-                # Continue with other servers even if one fails
-
-    async def _disconnect_mcp_servers(self):
-        """Disconnect from all configured MCP servers."""
-        if not self.mcp_servers:
-            return
-
-        self.logger.info(f"🔌 Disconnecting from {len(self.mcp_servers)} MCP server(s)")
-
-        for i, server in enumerate(self.mcp_servers):
-            try:
-                self.logger.info(
-                    f"  Disconnecting from MCP server {i + 1}/{len(self.mcp_servers)}: {server.__class__.__name__}"
-                )
-                await server.disconnect()
-                self.logger.info(
-                    f"  ✅ Disconnected from MCP server {i + 1}/{len(self.mcp_servers)}"
-                )
-            except Exception as e:
-                self.logger.error(
-                    f"  ❌ Error disconnecting from MCP server {i + 1}/{len(self.mcp_servers)}: {e}"
-                )
-                # Continue with other servers even if one fails
-
-    async def get_mcp_tools(self) -> List[Any]:
-        """Get all available tools from all connected MCP servers."""
-        tools = []
-
-        for server in self.mcp_servers:
-            if server.is_connected:
-                try:
-                    server_tools = await server.list_tools()
-                    tools.extend(server_tools)
-                except Exception as e:
-                    self.logger.error(
-                        f"Error getting tools from MCP server {server.__class__.__name__}: {e}"
-                    )
-
-        return tools
-
-    async def call_mcp_tool(
-        self, server_index: int, tool_name: str, arguments: Dict[str, Any]
-    ) -> Any:
-        """Call a tool on a specific MCP server.
-        Args:
-            server_index: Index of the MCP server in the mcp_servers list
-            tool_name: Name of the tool to call
-            arguments: Arguments to pass to the tool
-        Returns:
-            The result of the tool call
-        """
-        if server_index >= len(self.mcp_servers):
-            raise ValueError(f"Invalid server index: {server_index}")
-        server = self.mcp_servers[server_index]
-        if not server.is_connected:
-            raise RuntimeError(f"MCP server {server_index} is not connected")
-        return await server.call_tool(tool_name, arguments)
-
-    async def _register_mcp_tools(self, server_index: int, server: MCPBaseServer):
-        """Register tools from an MCP server with the LLM's function registry.
-
-        Args:
-            server_index: Index of the MCP server in the mcp_servers list
-            server: The connected MCP server
-        """
-        try:
-            # Get tools from the MCP server
-            mcp_tools = await server.list_tools()
-            self.logger.info(
-                f"  📋 Found {len(mcp_tools)} tools from MCP server {server_index + 1}"
-            )
-
-            # Register each tool with the function registry
-            for tool in mcp_tools:
-                try:
-                    # Create a wrapper function for the MCP tool
-                    tool_wrapper = MCPToolConverter.create_mcp_tool_wrapper(
-                        server_index, tool.name, self
-                    )
-
-                    # Register the tool with the LLM's function registry
-                    self.llm.function_registry.register(
-                        name=f"mcp_{server_index}_{tool.name}",
-                        description=tool.description or f"MCP tool: {tool.name}",
-                    )(tool_wrapper)
-
-                    self.logger.info(f"    ✅ Registered tool: {tool.name}")
-
-                except Exception as e:
-                    self.logger.error(
-                        f"    ❌ Failed to register tool {tool.name}: {e}"
-                    )
-                    # Continue with other tools even if one fails
-
-        except Exception as e:
-            self.logger.error(
-                f"  ❌ Failed to get tools from MCP server {server_index + 1}: {e}"
-            )
 
     def _truncate_for_logging(self, obj, max_length=200):
         """Truncate object string representation for logging to prevent spam."""
@@ -899,11 +773,3 @@ class Agent:
             obj_str = obj_str[:max_length] + "... (truncated)"
         return obj_str
 
-    def _setup_mcp_servers(self):
-        """Set up MCP servers if any are configured."""
-        if self.mcp_servers:
-            self.logger.info(f"🔌 Setting up {len(self.mcp_servers)} MCP server(s)")
-            for i, server in enumerate(self.mcp_servers):
-                self.logger.info(f"  {i + 1}. {server.__class__.__name__}")
-        else:
-            self.logger.debug("No MCP servers configured")
