+++ agents-core/stream_agents/core/mcp/__init__.py
@@ -4,5 +4,6 @@ from .mcp_server_remote import MCPServerRemote
 from .mcp_server_local import MCPServerLocal
 from .mcp_base import MCPBaseServer
 from .tool_converter import MCPToolConverter
+from .mcp_manager import MCPManager
 
-__all__ = ["MCPServerRemote", "MCPServerLocal", "MCPBaseServer", "MCPToolConverter"]
+__all__ = ["MCPServerRemote", "MCPServerLocal", "MCPBaseServer", "MCPToolConverter", "MCPManager"]
